+++ exercise08/ejercicio8.cpp
@@ -11,12 +11,10 @@ vector<int> closest_points(vector<vector<T>> const& points) {
 
     for (int i = 0; i < n; ++i) {
         for (int j = i + 1; j < n; ++j) {
-            // Calculamos la distancia al cuadrado entre el punto i y el punto j
-            double dx = points[j][0] - points[i][0];
-            double dy = points[j][1] - points[i][1];
-            double distance_sq = dx * dx + dy * dy;
+            long double dx = points[j][0] - points[i][0];
+            long double dy = points[j][1] - points[i][1];
+            long double distance_sq = dx * dx + dy * dy;
 
-            // Si encontramos una distancia menor, actualizamos el par de puntos más cercanos
             if (distance_sq < min_distance) {
                 min_distance = distance_sq;
                 result = {i, j};
