+++ exercise10/ejercicio10.cpp
@@ -3,8 +3,8 @@ using namespace std;
 
 const int MOD = 1000000007;
 
-vector<int> numeros_catalan(int k) {
-    vector<int> catalan(k + 1, 0);
+vector<long int> numeros_catalan(int k) {
+    vector<long int> catalan(k + 1, 0);
     catalan[0] = 1;  // C_0 es 1
 
     for (int i = 1; i <= k; ++i) {
@@ -21,7 +21,7 @@ int num_of_triangulations(int n) {
         return 0;  // No es posible hacer triangulaciones en polígonos con menos de 3 lados
     }
 
-    vector<int> catalan = numeros_catalan(n - 2);
+    vector<long int> catalan = numeros_catalan(n - 2);
 
     return catalan[n - 2];
 }
